правочка

+++ src/TenClose.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class TenClose {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int m = sc.nextInt();
-        int n = sc.nextInt();
+        float m = sc.nextFloat();
+        float n = sc.nextFloat();
         if (m - 10 > n - 10) {
             System.out.println(n);
         } else {
